update: added offline crud to subTodos

+++ src/features/subTodos/subTodosSlice.js
@@ -5,8 +5,8 @@ import { baseUrl } from "../../app/shared/baseUrl";
 export const fetchSubTodos = createAsyncThunk(
   "subTodos/fetchSubTodos",
   async () => {
-    // item is in double quotes when retrieved ""item""
-    const user = localStorage.getItem("user").replace(/['"]+/g, "");
+    const user = JSON.parse(localStorage.getItem("user"));
+
     if (user === "offline") {
       return JSON.parse(localStorage.getItem("subTodos") ?? "[]");
     } else {
@@ -31,34 +31,46 @@ export const postSubTodo = createAsyncThunk(
       return Promise.reject("empty input");
     }
 
-    const response = await fetch(baseUrl + "subTodos", {
-      method: "POST",
-      body: JSON.stringify(subTodo),
-      headers: { "Content-Type": "application/json" },
-      credentials: "include",
-    });
+    const user = JSON.parse(localStorage.getItem("user"));
 
-    if (!response.ok) {
-      return Promise.reject(response.status);
+    if (user === "offline") {
+      dispatch(addSubTodo(subTodo));
+    } else {
+      const response = await fetch(baseUrl + "subTodos", {
+        method: "POST",
+        body: JSON.stringify(subTodo),
+        headers: { "Content-Type": "application/json" },
+        credentials: "include",
+      });
+
+      if (!response.ok) {
+        return Promise.reject(response.status);
+      }
+
+      const data = await response.json();
+      dispatch(addSubTodo(data));
     }
-
-    const data = await response.json();
-    dispatch(addSubTodo(data));
   },
 );
 
 export const updateSubTodo = createAsyncThunk(
   "subTodos/updateSubTodo",
   async (subTodo) => {
-    const response = await fetch(baseUrl + `subTodos/${subTodo._id}`, {
-      method: "PUT",
-      body: JSON.stringify(subTodo),
-      headers: { "Content-Type": "application/json" },
-      credentials: "include",
-    });
+    const user = JSON.parse(localStorage.getItem("user"));
 
-    if (!response.ok) {
-      return Promise.reject(response.status);
+    if (user === "offline") {
+      return subTodo;
+    } else {
+      const response = await fetch(baseUrl + `subTodos/${subTodo._id}`, {
+        method: "PUT",
+        body: JSON.stringify(subTodo),
+        headers: { "Content-Type": "application/json" },
+        credentials: "include",
+      });
+
+      if (!response.ok) {
+        return Promise.reject(response.status);
+      }
     }
   },
 );
@@ -66,17 +78,23 @@ export const updateSubTodo = createAsyncThunk(
 export const deleteSubTodo = createAsyncThunk(
   "subTodos/deleteSubTodo",
   async (subTodo) => {
-    const response = await fetch(baseUrl + `subTodos/${subTodo._id}`, {
-      method: "DELETE",
-      credentials: "include",
-    });
+    const user = JSON.parse(localStorage.getItem("user"));
 
-    if (!response.ok) {
-      return Promise.reject(response.status);
-    }
+    if (user === "offline") {
+      return subTodo;
+    } else {
+      const response = await fetch(baseUrl + `subTodos/${subTodo._id}`, {
+        method: "DELETE",
+        credentials: "include",
+      });
 
-    if (response.ok) {
-      return { _id: subTodo._id };
+      if (!response.ok) {
+        return Promise.reject(response.status);
+      }
+
+      if (response.ok) {
+        return { _id: subTodo._id };
+      }
     }
   },
 );
@@ -111,10 +129,24 @@ const subTodosSlice = createSlice({
   initialState,
   reducers: {
     addSubTodo: (state, action) => {
-      const newSubTodo = {
-        ...action.payload,
-      };
-      state.subTodosArray.push(newSubTodo);
+      const user = JSON.parse(localStorage.getItem("user"));
+
+      if (user === "offline") {
+        const newSubTodo = {
+          _id: state.subTodosArray.length + 1,
+          ...action.payload,
+        };
+
+        state.subTodosArray.push(newSubTodo);
+        let localArr = JSON.stringify(state.subTodosArray);
+
+        localStorage.setItem("subTodos", localArr);
+      } else {
+        const newSubTodo = {
+          ...action.payload,
+        };
+        state.subTodosArray.push(newSubTodo);
+      }
     },
   },
   extraReducers: {
@@ -132,20 +164,52 @@ const subTodosSlice = createSlice({
     },
     [postSubTodo.rejected]: (action) => {
       alert(
-        "Your todo could not be posted\nError: " +
+        "Your sub todo could not be posted\nError: " +
           (action.error
             ? action.error.message
             : "Please provide a valid input"),
       );
     },
     [deleteSubTodo.fulfilled]: (state, action) => {
-      const deletedSubTodoId = action.payload._id;
+      const user = JSON.parse(localStorage.getItem("user"));
 
-      const updatedSubTodosArray = state.subTodosArray.filter(
-        (subTodo) => subTodo._id !== deletedSubTodoId,
-      );
+      if (user === "offline") {
+        let subTodos = JSON.parse(localStorage.getItem("subTodos"));
 
-      state.subTodosArray = updatedSubTodosArray;
+        let filtered = subTodos.filter(
+          (subTodo) => subTodo._id !== action.payload._id,
+        );
+        state.subTodosArray = filtered;
+
+        let stringArr = JSON.stringify(filtered);
+        localStorage.setItem("subTodos", stringArr);
+      } else {
+        const deletedSubTodoId = action.payload._id;
+
+        const updatedSubTodosArray = state.subTodosArray.filter(
+          (subTodo) => subTodo._id !== deletedSubTodoId,
+        );
+
+        state.subTodosArray = updatedSubTodosArray;
+      }
+    },
+    [updateSubTodo.fulfilled]: (state, action) => {
+      const user = JSON.parse(localStorage.getItem("user"));
+
+      if (user === "offline") {
+        let subTodos = JSON.parse(localStorage.getItem("subTodos"));
+
+        const newSubTodos = subTodos.map((subTodo) => {
+          if (subTodo._id === action.payload._id) {
+            return { ...subTodo, subText: action.payload.subText };
+          }
+          return subTodo;
+        });
+
+        let result = JSON.stringify(newSubTodos);
+        localStorage.setItem("subTodos", result);
+        state.subTodosArray = newSubTodos;
+      }
     },
     [updateSubTodoDone.fulfilled]: (state, action) => {
       const updateSubTododoneID = action.payload._id;
